split layer interfaces and implementations into separate files
